+++ src/main/java/forestry/apiculture/gui/GuiAlveary.java
@@ -10,24 +10,27 @@
  ******************************************************************************/
 package forestry.apiculture.gui;
 
-import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.network.chat.Component;
+import net.minecraft.world.entity.player.Inventory;
+import net.minecraft.world.inventory.Slot;
+import net.minecraft.world.item.ItemStack;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 
 import forestry.apiculture.multiblock.IAlvearyControllerInternal;
 import forestry.apiculture.multiblock.TileAlveary;
 import forestry.core.config.Constants;
-import forestry.core.gui.GuiForestryTitled;
+import forestry.core.gui.GuiAnalyzerProvider;
 import forestry.core.render.EnumTankLevel;
 
-public class GuiAlveary extends GuiForestryTitled<ContainerAlveary> {
+public class GuiAlveary extends GuiAnalyzerProvider<ContainerAlveary> {
 	private final TileAlveary tile;
 
 	public GuiAlveary(ContainerAlveary container, Inventory inventory, Component title) {
-		super(Constants.TEXTURE_PATH_GUI + "/alveary.png", container, inventory, title);
+		super(Constants.TEXTURE_PATH_GUI + "/alveary.png", container, inventory, () -> title, 25, 7, 2, 0);
 		this.tile = container.getTile();
 		this.imageHeight = 190;
+		analyzer.init();
 	}
 
 	@Override
@@ -51,5 +54,22 @@ public class GuiAlveary extends GuiForestryTitled<ContainerAlveary> {
 		addClimateLedger(tile);
 		addHintLedger("apiary");
 		addOwnerLedger(tile);
+	}
+
+	@Override
+	protected void drawSelectedSlot(PoseStack transform, int selectedSlot) {
+		Slot slot = menu.getForestrySlot(1 + selectedSlot);
+		SELECTED_COMB_SLOT.draw(transform, topPos + slot.y - 3, leftPos + slot.x - 3);
+	}
+
+	@Override
+	public ItemStack getSpecimen(int index) {
+		Slot slot = menu.getForestrySlot(getSelectedSlot(index));
+		return slot.getItem();
+	}
+
+	@Override
+	protected boolean hasErrors() {
+		return tile.getErrorLogic().hasErrors();
 	}
 }