+++ src/main/java/forestry/core/tiles/ITitled.java
@@ -1,6 +1,7 @@
 package forestry.core.tiles;
 
-//TODO
-public interface ITitled {    //TODO perhaps make this return an ITextComponent for ease of use with INamedContainerProvider.
-	String getUnlocalizedTitle();
+import net.minecraft.network.chat.Component;
+
+public interface ITitled {
+	Component getTitle();
 }